packages/dopetracks-mcp/pyproject.toml
[project]
name = "dopetracks-mcp"
version = "0.1.0"
description = "MCP server for Dopetracks - iMessage Spotify playlist creation"
authors = [{name = "namarks", email = "nmarkspdx@gmail.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "spotipy>=2.23.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "typedstream>=0.0.6",
    "requests>=2.28.0",
    "numpy>=1.24.0",
]

[project.scripts]
dopetracks-mcp = "dopetracks_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dopetracks_mcp"]
